fix: tests should run on Mac

+++ test/run-tests.js
@@ -26,13 +26,11 @@ function awaitElectron() {
   cleanupElectron();
 }
 
-if (!isCI) {
-  shell.echo('\n*** Reinstalling from scratch\n');
-  if (shell.test('-d', 'node_modules')) {
-    shell.rm('-rf', 'node_modules');
-  }
-  shell.exec('npm install');
+shell.echo('\n*** Reinstalling from scratch\n');
+if (!isCI && shell.test('-d', 'node_modules')) {
+  shell.rm('-rf', 'node_modules');
 }
+shell.exec('npm install');
 
 shell.echo('\n*** Running simple tests\n');
 if (shell.test('-f', 'path.txt')) {
@@ -46,18 +44,23 @@ cleanupElectron();
 shell.echo('\n*** Running packaged test\n');
 shell.exec('npm run pack');
 var executable;
+var unpackedPath = 'resources/app.asar.unpacked';
 if (shell.test('-d', 'dist/linux-unpacked')) {
   shell.cd('dist/linux-unpacked');
   executable = './electron-prebuilt-path-test';
 } else if (shell.test('-d', 'dist/win-unpacked')) {
   shell.cd('dist/win-unpacked');
   executable = 'electron-prebuilt-path-test.exe';
+} else if (shell.test('-d', 'dist/mac')) {
+  shell.cd('dist/mac');
+  executable = 'electron-prebuilt-path-test.app/Contents/MacOS/electron-prebuilt-path-test';
+  unpackedPath = 'electron-prebuilt-path-test.app/Contents/Resources/app.asar.unpacked';
 }
-shell.cp('../../get-path.js', 'resources/app.asar.unpacked/');
-shell.cd('resources/app.asar.unpacked/');
+shell.cp('../../get-path.js', unpackedPath);
+shell.pushd(unpackedPath);
 shell.exec('node get-path.js');
-shell.mv('path.txt', '../../');
-shell.cd('../../');
+shell.popd();
+shell.mv(unpackedPath + '/path.txt', '.');
 shell.exec(executable);
 awaitElectron();
 shell.rm('-rf', 'dist');
